+++ indispensable-cosef/src/jvmTest/kotlin/ConversionTests.kt
@@ -4,8 +4,10 @@ import at.asitplus.signum.indispensable.cosef.toCoseAlgorithm
 import at.asitplus.signum.indispensable.cosef.toCoseKey
 import at.asitplus.signum.indispensable.toCryptoPublicKey
 import at.asitplus.signum.indispensable.toX509SignatureAlgorithm
-import io.kotest.core.spec.style.FreeSpec
-import io.kotest.datatest.withData
+import at.asitplus.testballoon.invoke
+import at.asitplus.testballoon.minus
+import at.asitplus.testballoon.withData
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.matchers.shouldBe
 import java.security.KeyPairGenerator
 import java.security.interfaces.ECPublicKey
@@ -15,7 +17,7 @@ import kotlin.random.Random
 infix fun <T> KmmResult<T>.shouldSucceedWith(b: T): T =
     (this.getOrThrow() shouldBe b)
 
-class ConversionTests : FreeSpec({
+val ConversionTests by testSuite {
     "COSE -> SigAlg -> COSE is stable" - {
 
         "All" - {
@@ -45,7 +47,7 @@ class ConversionTests : FreeSpec({
         val kid = Random.nextBytes(16)
         cpk.toCoseKey(keyId = kid).getOrThrow().keyId shouldBe kid
     }
-})
+}
 
 private fun randomPublicKey() =
     (KeyPairGenerator.getInstance("EC").apply { initialize(256) }