+++ indispensable/src/commonTest/kotlin/PemTest.kt
@@ -6,13 +6,14 @@ import at.asitplus.signum.indispensable.asn1.encodeToPEM
 import at.asitplus.signum.indispensable.pki.Pkcs10CertificationRequest
 import at.asitplus.signum.indispensable.pki.X509Certificate
 
-import io.kotest.core.spec.style.FreeSpec
+import at.asitplus.testballoon.invoke
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlin.random.Random
 
 @OptIn(ExperimentalStdlibApi::class)
-class PemTest : FreeSpec({
+val PemTest  by testSuite{
 
    
     "Cert"  {
@@ -498,4 +499,4 @@ class PemTest : FreeSpec({
             CryptoPrivateKey.decodeFromPem(it).getOrThrow()
         }
     }
-})
+}