+++ supreme/src/commonTest/kotlin/ProjectConfig.kt
@@ -1,6 +1,0 @@
-package io.kotest.provided
-
-import at.asitplus.test.XmlReportingProjectConfig
-
-/** Wires KMP JUnit XML reporting */
-class ProjectConfig : XmlReportingProjectConfig()