+++ supreme/src/commonTest/kotlin/at/asitplus/signum/supreme/mac/MACTest.kt
@@ -5,10 +5,16 @@ import at.asitplus.signum.indispensable.HMAC
 import at.asitplus.signum.indispensable.misc.bit
 import at.asitplus.signum.supreme.b
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.datatest.withData
+import at.asitplus.testballoon.minus
+import at.asitplus.testballoon.invoke
+import at.asitplus.testballoon.withData
+import at.asitplus.testballoon.withDataSuites
+import at.asitplus.testballoon.checkAllTests
+import at.asitplus.testballoon.checkAllSuites
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.matchers.shouldBe
 
-class MACTest : FreeSpec({
+val MACTest  by testSuite{
     "RFC4231" - {
         class I(val comment: String, key: String, data: String,
                 SHA224: String, SHA256: String, SHA384: String, SHA512: String) {
@@ -94,4 +100,4 @@ class MACTest : FreeSpec({
             .mac(key=b("0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b"), b("4869205468657265"))
             .getOrThrow() shouldBe byteArrayOf(0xa0.toByte())
     }
-})
+}