+++ supreme/src/commonTest/kotlin/at/asitplus/signum/supreme/sign/PrivateKeyCommonTests.kt
@@ -5,12 +5,18 @@ import at.asitplus.signum.indispensable.SignatureAlgorithm
 import at.asitplus.signum.indispensable.SecretExposure
 import at.asitplus.signum.supreme.isSuccess
 import at.asitplus.signum.supreme.signature
-import io.kotest.core.spec.style.FreeSpec
+import at.asitplus.testballoon.minus
+import at.asitplus.testballoon.invoke
+import at.asitplus.testballoon.withData
+import at.asitplus.testballoon.withDataSuites
+import at.asitplus.testballoon.checkAllTests
+import at.asitplus.testballoon.checkAllSuites
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.matchers.shouldBe
 import kotlin.random.Random
 
 @OptIn(SecretExposure::class)
-class PrivateKeyCommonTests : FreeSpec({
+val PrivateKeyCommonTests  by testSuite{
     "RSA" {
         val rsa = """
             -----BEGIN PRIVATE KEY-----
@@ -97,4 +103,4 @@ class PrivateKeyCommonTests : FreeSpec({
             CryptoPrivateKey.decodeFromDer("3041020100301306072a8648ce3d020106082a8648ce3d03010704273025020101042001811d2b378be969f614283650e8ca3b07eba2289841239513e24fd230e5a538".hexToByteArray())
         }
     }
-})
+}