+++ supreme/src/iosTest/kotlin/Test.kt
@@ -1,11 +1,13 @@
 package at.asitplus.signum.supreme
-import io.kotest.core.spec.style.FreeSpec
+
+import at.asitplus.testballoon.invoke
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.matchers.shouldNotBe
 
-class ProviderTest : FreeSpec({
+val ProviderTest by testSuite {
 
     "This dummy test" {
         "is just making sure" shouldNotBe "that iOS tests are indeed running"
     }
 
-})
+}