+++ supreme/src/jvmTest/kotlin/at/asitplus/signum/supreme/sign/VerifierTests.kt
@@ -4,7 +4,13 @@ import at.asitplus.catching
 import at.asitplus.signum.indispensable.*
 import at.asitplus.signum.supreme.succeed
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.datatest.withData
+import at.asitplus.testballoon.minus
+import at.asitplus.testballoon.invoke
+import at.asitplus.testballoon.withData
+import at.asitplus.testballoon.withDataSuites
+import at.asitplus.testballoon.checkAllTests
+import at.asitplus.testballoon.checkAllSuites
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.matchers.should
 import io.kotest.matchers.shouldNot
 import io.kotest.matchers.types.shouldBeInstanceOf
@@ -15,15 +21,24 @@ import java.security.Signature
 import java.security.spec.ECGenParameterSpec
 import kotlin.random.Random
 
-class VerifierTests: FreeSpec({
-    withData(mapOf<String, (SignatureAlgorithm.ECDSA, CryptoPublicKey.EC)-> Verifier.EC>(
-        "BC -> PlatformVerifier" to { a,k ->
-            a.verifierFor(k) { provider = "BC" }.getOrThrow().also { it.shouldBeInstanceOf<PlatformECDSAVerifier>() }
-        },
-        "BC -> KotlinVerifier" to ::KotlinECDSAVerifier)) { factory ->
-        withData(ECCurve.entries) { curve ->
-            withData(nameFn = SignatureInputFormat::jcaAlgorithmComponent, listOf<Digest?>(null) + Digest.entries) { digest ->
-                withData(nameFn = { (key,_,_) -> key.publicPoint.toString() }, generateSequence {
+val VerifierTests by testSuite {
+    Security.addProvider(BouncyCastleProvider())
+
+    withDataSuites(
+        mapOf<String, (SignatureAlgorithm.ECDSA, CryptoPublicKey.EC) -> Verifier.EC>(
+            "BC -> PlatformVerifier" to { a, k ->
+                a.verifierFor(k) { provider = "BC" }.getOrThrow()
+                    .also { it.shouldBeInstanceOf<PlatformECDSAVerifier>() }
+            },
+            "BC -> KotlinVerifier" to ::KotlinECDSAVerifier
+        )
+    ) { factory ->
+        withDataSuites(ECCurve.entries)  { curve ->
+            withDataSuites(
+                nameFn = SignatureInputFormat::jcaAlgorithmComponent,
+                listOf<Digest?>(null) + Digest.entries
+            ) { digest ->
+                withData(nameFn = { (key, _, _) -> key.publicPoint.toString() }, generateSequence {
                     val keypair = KeyPairGenerator.getInstance("EC", "BC").also {
                         it.initialize(ECGenParameterSpec(curve.jcaName))
                     }.genKeyPair()
@@ -52,4 +67,4 @@ class VerifierTests: FreeSpec({
             }
         }
     }
-}) { companion object { init { Security.addProvider(BouncyCastleProvider())}}}
+}